Have the Adversary choose best seller
Instead of having F_SAT choosing the best seller, it creates a list of
possible sellers, passes it on to the adversary and expects it to choose
one of them. This way some relaxation from the optimal F_SAT is given.
Note that this is the first appearance of the Adversary.