Have F_SAT choose the best seller if A fails
In case the Adversary does not give a valid response when F_SAT sends
the list of potential choices (e.g. if A responds with seller = \bot),
then F_SAT chooses the best seller, price and asset from the list. This
way the Adversary cannot entirely block trades from completing.